[chore] - 여정화면에서 탭바 삭제
- settingDeviceToken()은 SessionStartViewController로 이동
