+++ Carmu/Sources/Controllers/SessionStart/SessionStartViewController.swift
@@ -435,3 +435,19 @@ extension SessionStartViewController: UICollectionViewDataSource {
         return cell ?? UICollectionViewCell()
     }
 }
+
+extension SessionStartViewController {
+
+    // 디바이스 토큰값이 업데이트가 될 수도 있기 때문에 업데이트된 값을 저장하기 위한 메서드
+    private func settingDeviceToken() {
+        print("settingDeviceToken()")
+        guard let databasePath = User.databasePathWithUID else {
+            return
+        }
+        guard let fcmToken = KeychainItem.currentUserDeviceToken else {
+            return
+        }
+        print("FCMToken -> ", fcmToken)
+        databasePath.child("deviceToken").setValue(fcmToken as NSString)
+    }
+}
